created localQuickActionCard
