+++ litmus-portal/frontend/src/views/MyHub/MyHubConnect/index.tsx
@@ -17,6 +17,7 @@ import ButtonOutline from '../../../components/Button/ButtonOutline';
 import GithubInputFields from '../../../components/GitHubComponents/GithubInputFields/GithubInputFields';
 import GitHubToggleButton from '../../../components/GitHubComponents/GitHubToggleButtons/GitHubToggleButton';
 import Loader from '../../../components/Loader';
+import { LocalQuickActionCard } from '../../../components/LocalQuickActionCard';
 import VideoCarousel from '../../../components/VideoCarousel';
 import Scaffold from '../../../containers/layouts/Scaffold';
 import {
@@ -31,7 +32,6 @@ import {
   SSHKey,
   SSHKeys,
 } from '../../../models/graphql/user';
-import { HomePageQuickActionCard } from '../../../pages/HomePage/homePageQuickActionCard';
 import { history } from '../../../redux/configureStore';
 import { RootState } from '../../../redux/reducers';
 import { validateStartEmptySpacing } from '../../../utils/validate';
@@ -549,7 +549,7 @@ const MyHub = () => {
             {t('myhub.connectHubPage.videoDesc')}
           </Typography>
           <div className={classes.quickActionDiv}>
-            <HomePageQuickActionCard />
+            <LocalQuickActionCard variant="homePage" />
           </div>
         </div>
       </div>
