+++ litmus-portal/frontend/src/views/MyHub/MyHubEdit/index.tsx
@@ -18,6 +18,7 @@ import ButtonOutline from '../../../components/Button/ButtonOutline';
 import GithubInputFields from '../../../components/GitHubComponents/GithubInputFields/GithubInputFields';
 import GitHubToggleButton from '../../../components/GitHubComponents/GitHubToggleButtons/GitHubToggleButton';
 import Loader from '../../../components/Loader';
+import { LocalQuickActionCard } from '../../../components/LocalQuickActionCard';
 import VideoCarousel from '../../../components/VideoCarousel';
 import Scaffold from '../../../containers/layouts/Scaffold';
 import { GENERATE_SSH, GET_HUB_STATUS, UPDATE_MY_HUB } from '../../../graphql';
@@ -29,7 +30,6 @@ import {
   SSHKeys,
 } from '../../../models/graphql/user';
 import { HubStatus } from '../../../models/redux/myhub';
-import { HomePageQuickActionCard } from '../../../pages/HomePage/homePageQuickActionCard';
 import { history } from '../../../redux/configureStore';
 import { RootState } from '../../../redux/reducers';
 import { validateStartEmptySpacing } from '../../../utils/validate';
@@ -488,7 +488,7 @@ const MyHub = () => {
             {t('myhub.connectHubPage.videoDesc')}
           </Typography>
           <div className={classes.quickActionDiv}>
-            <HomePageQuickActionCard />
+            <LocalQuickActionCard variant="homePage" />
           </div>
         </div>
       </div>
